fix(dashboard): cast precisa_revisao to IntegerField in aggregation
Update the `DashboardView` to cast `precisa_revisao` to an `IntegerField`
instead of a `FloatField` during aggregation to ensure correct data
type handling for the average calculation.

+++ apps/dashboard/views.py
@@ -9,9 +9,8 @@ from datetime import timedelta
 import json
 
 from django.contrib.auth.mixins import LoginRequiredMixin
-from django.db.models import Avg, Count, F, Q
+from django.db.models import Avg, Count, F, IntegerField, Q
 from django.db.models.functions import Cast
-from django.db.models import FloatField
 from django.http import HttpResponse
 from django.utils import timezone
 from django.views.generic import TemplateView
@@ -68,7 +67,7 @@ class DashboardView(LoginRequiredMixin, TemplateView):
         llm_stats = DadosExtraidos.objects.aggregate(
             total_extracoes=Count("id"),
             taxa_precisa_revisao=Avg(
-                Cast("precisa_revisao", output_field=FloatField())
+                Cast("precisa_revisao", output_field=IntegerField())
             ),
             tempo_medio_ms=Avg("tempo_resposta_ms"),
         )
